requirements.txt
numpy
pandas
matplotlib
opencv-python
streamlit
streamlit-option-menu
PyYAML